Add new Cash system

+++ AmericanChess/source/GameObject/ChessObject/ChessPiece.cpp
@@ -45,7 +45,7 @@ void ChessPiece::init(ChessPosition pos) {
     if (m_type == PLAYER) return;
     m_health = GameRule->getHealthChess(m_type);
     m_queueSize = GameRule->getQueueSizeChess(m_type);
-    m_turnLeft = reinterpret_cast<int>(this) % m_queueSize + 1;
+    m_turnLeft = reinterpret_cast<long long>(this) % m_queueSize + 1;
 }
 
 void ChessPiece::init(ChessPosition pos, int health, int turnLeft, int queueSize) {
@@ -62,7 +62,7 @@ void ChessPiece::update(float deltaTime) {
     case IDLE:
         if (m_isPromotion) {
             std::vector<std::string> promoteList = { "W_Queen", "W_Knight", "W_Rook", "W_Bishop" };
-            promote(promoteList[reinterpret_cast<int>(this) % promoteList.size()]);
+            promote(promoteList[reinterpret_cast<long long>(this) % promoteList.size()]);
         }
         break;
     case READY_TO_MOVE:
